+++ test/Scripts.Integration.Test/integration-test.ps1
@@ -1,4 +1,9 @@
-﻿param(
+# ┌───────────────────────────────────────────────────┐ #	
+# │    This script is for local use only,       	  │ #
+# │    utilizing the scripts locally we use in CI.    │ #
+# └───────────────────────────────────────────────────┘ #	
+
+param(
 	[string] $UnityVersion,
 	[string] $Platform,
 	[switch] $Clean,