+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Automated Performance Instrumentalisation for Runtime Initialization ([#991](https://github.com/getsentry/sentry-unity/pull/991))
 - Automated Performance Instrumentalisation for Scene Loading ([#768](https://github.com/getsentry/sentry-unity/pull/768))
+- Automated Performance Instrumentation for Scene Loading ([#768](https://github.com/getsentry/sentry-unity/pull/768))
 
 ### Dependencies